+++ src/devicecontroller/DeviceSelection.ts
@@ -5,6 +5,7 @@ export default class DeviceSelection {
   constraints: MediaStreamConstraints;
   stream: MediaStream;
   groupId: string = '';
+  endedCallback?: undefined | (() => void);
 
   matchesConstraints(constraints: MediaStreamConstraints): boolean {
     return JSON.stringify(this.constraints) === JSON.stringify(constraints);