+++ src/pingpong/DefaultPingPong.ts
@@ -48,7 +48,7 @@ export default class DefaultPingPong implements SignalingClientObserver, PingPon
 
   forEachObserver(observerFunc: (observer: PingPongObserver) => void): void {
     for (const observer of this.observerQueue) {
-      new AsyncScheduler().start(() => {
+      AsyncScheduler.nextTick(() => {
         if (this.observerQueue.has(observer)) {
           observerFunc(observer);
         }