+++ src/scheduler/AsyncScheduler.ts
@@ -11,4 +11,20 @@ export default class AsyncScheduler extends TimeoutScheduler {
   constructor() {
     super(0);
   }
+
+  /**
+   * Execute the provided callback on the next tick of the event loop.
+   * This is semantically equivalent to
+   *
+   * ```typescript
+   * new AsyncScheduler(callback).start();
+   * ```
+   *
+   * but with less overhead.
+   *
+   * @param callback the code to run.
+   */
+  static nextTick(callback: () => void): void {
+    setTimeout(callback, 0);
+  }
 }